requirements.txt
cachelib==0.9.0
click==8.1.3
colorama==0.4.5
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-Session==0.4.0
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.23.2
opencv-python==4.6.0.66
Pillow==9.2.0
qrcode==7.3.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.1